feat: update the api
- update on repository customer and province.
  for customer added new function
  for province fix name and add where condition
- fix some typo and data type on models
- add interface payment installment
- add implementation interface repositories (dail_loan_request, customer_loan_request and payment_installment) on gorm packages.

+++ httpserver/repositories/models/customer_loan_requests.go
@@ -6,7 +6,7 @@ type CustomerLoanRequest struct {
 	Id         uint
 	CustomerId uint
 	Amount     uint
-	Tenor      int8
+	Tenor      uint
 	Status     string
 	CreatedAt  time.Time
 	UpdatedAt  time.Time
